Refactor CF client creation out of validate func
In order to determine org/app/space names we will add new functions
accessing the CF client. To minimise the amount of calls to NewCFClient,
the call is moved out of the validate function and the client is passed
to validate instead of the config. Typing requires to add go-cfclient as
a dependency, which was indirectly present via util before.