Add CF org/app/space names to Auth.Alias.Metadata (#39)
* Test CF org/space/app names in Auth.Alias.Metadata

Metadata creation is tested as part of the backend tests. Also updating
the mock cf server to match current testing pattern.

* Refactor CF client creation out of validate func

In order to determine org/app/space names we will add new functions
accessing the CF client. To minimise the amount of calls to NewCFClient,
the call is moved out of the validate function and the client is passed
to validate instead of the config. Typing requires to add go-cfclient as
a dependency, which was indirectly present via util before.

* Add CF org/app/space name to Auth.Alias.Metadata

This allows to set policies based on the org, app and space names, rather
than GUIDs. Doing so is helpful if the GUIDs for a given name change,
e.g. during blue/green deployment of an app. It also makes it easier to
identify a secret's context.